requirements.txt
aiogram
openai
requests
python-dotenv
sqlite3
